+++ spec/rubocop/ast/break_node_spec.rb
@@ -1,13 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'wrapped_arguments_node'
+
 RSpec.describe RuboCop::AST::BreakNode do
-  let(:break_node) { parse_source(source).ast }
-
-  describe '.new' do
-    context 'with a break node' do
-      let(:source) { 'break' }
-
-      it { expect(break_node.is_a?(described_class)).to be(true) }
-    end
-  end
+  it_behaves_like 'wrapped arguments node', 'break'
 end