+++ lib/rubocop/ast/node/return_node.rb
@@ -6,7 +6,6 @@ module RuboCop
     # plain node when the builder constructs the AST, making its methods
     # available to all `return` nodes within RuboCop.
     class ReturnNode < Node
-      include MethodDispatchNode
       include ParameterizedNode
       include WrappedArgumentsNode
     end